fix logic prev push
