+++ src/view/gui/AsyncExtract.h
@@ -1,75 +1,68 @@
-// #ifndef ASYNCEXTRACT_H
-// #define ASYNCEXTRACT_H
+#ifndef ASYNCEXTRACT_H
+#define ASYNCEXTRACT_H
 
-// #include <QObject>
-// #include <QThread>
-// #include <QFile>
+#include <QObject>
+#include <QThread>
+#include <QFile>
 
-// #include "CaPTkUtils.h"
-// #include "CaPTkGUIUtils.h"
-// #include "QZipReader.h"
-// #include "ApplicationPreferences.h"
+#include "CaPTkUtils.h"
+#include "CaPTkGUIUtils.h"
+#include "QZipReader.h"
+#include "ApplicationPreferences.h"
 
-// // class ASyncExtract : public QThread
-// // {
-// // 	Q_OBJECT
-// // 	void run() override {
-// // 		ApplicationPreferences::GetInstance()->SetLibraExtractionStartedStatus(QVariant("true").toString());
+class ASyncExtract : public QThread
+{
+	Q_OBJECT
+	void run() override {
+		ApplicationPreferences::GetInstance()->SetLibraExtractionStartedStatus(QVariant("true").toString());
+		ApplicationPreferences::GetInstance()->SerializePreferences();
+		ApplicationPreferences::GetInstance()->DisplayPreferences();
 
-// // 		if (QFile::exists(this->fullPath))
-// // 		{
-// // 			QZipReader zr(this->fullPath);
-// // 			bool ret = zr.extractAll(this->extractPath);
+		if (QFile::exists(this->fullPath))
+		{
+			QZipReader zr(this->fullPath);
+			bool ret = zr.extractAll(this->extractPath);
 
-// // 			if(ret)
-// // 			{
-// // 				ApplicationPreferences::GetInstance()->SetLibraExtractionFinishedStatus(QVariant("true").toString());
-// // 				//after extraction remove the zip
-// // 				bool successfullyremoved = QFile::remove(this->fullPath.toStdString().c_str());
-// // 			}
-// // 			else
-// // 			{
-// // 				ApplicationPreferences::GetInstance()->SetLibraExtractionFinishedStatus(QVariant("false").toString());
-// // 			}
-// // 		}
-// // 		else
-// // 		{
-// // 			ApplicationPreferences::GetInstance()->SetLibraExtractionFinishedStatus(QVariant("false").toString());
-// // 		}
+			if(ret)
+			{
+				//after extraction remove the zip
+				bool successfullyremoved = QFile::remove(this->fullPath.toStdString().c_str());
+			}
+		}
 
-// // 		qDebug() << "Extraction done in background" << this->fullPath << endl;
+		qDebug() << "Extraction done in background" << this->fullPath << endl;
 
-// // 		//serialize only once
-// // 		ApplicationPreferences::GetInstance()->SerializePreferences();
+		//serialize only once
+		ApplicationPreferences::GetInstance()->SerializePreferences();
 
-// // 		emit resultReady(this->appName);
-// // 	}	
+		emit resultReady(this->appName);
+	}	
 
-// // public:
-// // 	ASyncExtract() = default;
-// // 	~ASyncExtract() = default;
+public:
+	ASyncExtract() = default;
+	~ASyncExtract() = default;
 
-// // 	void setFullPath(QString fullPath) {
-// // 		this->fullPath = fullPath;
-// // 	}
+	void setFullPath(QString fullPath) {
+		this->fullPath = fullPath;
+	}
 
-// // 	void setExtractPath(QString extractPath) {
-// // 		this->extractPath = extractPath;
-// // 	}
+	void setExtractPath(QString extractPath) {
+		this->extractPath = extractPath;
+	}
 
-// // 	void setAppName(QString appName) {
-// // 		this->appName = appName;
-// // 	}
+	void setAppName(QString appName) {
+		this->appName = appName;
+	}
 
-// // private:
-// // 	Q_DISABLE_COPY(ASyncExtract)
+private:
+	Q_DISABLE_COPY(ASyncExtract)
 
-// // 	QString fullPath;
-// // 	QString extractPath;
-// // 	QString appName;
+	QString fullPath;
+	QString extractPath;
+	QString appName;
 
-// // signals:
-// //     void resultReady(QString appName);
-// // };
+signals:
+    void resultReady(QString appName);
+};
 
-// #endif
+#endif
