+++ src/view/gui/StandaloneApp.cpp
@@ -10,21 +10,6 @@
 #include "CaPTkGUIUtils.h"
 #include "ApplicationPreferences.h"
 
-// StandaloneApp* StandaloneApp::m_Instance = nullptr;
-// QMutex StandaloneApp::m_Mutex;
-
-// StandaloneApp* StandaloneApp::GetInstance()
-// {
-// 	if (m_Instance == nullptr)
-// 	{
-// 		m_Mutex.lock();
-// 		m_Instance = new StandaloneApp();
-// 		// m_Instance.SetName(appName);
-// 		m_Mutex.unlock();
-// 	}
-// 	return m_Instance;
-// }
-
 void StandaloneApp::SetName(QString appName)
 {
 	this->m_AppName = appName;
@@ -36,48 +21,32 @@ QString StandaloneApp::GetName() const
 }
 
 std::string StandaloneApp::getStandaloneApp(QString appName) {
-	// this = pMainWindow;
 	this->m_AppName = appName;
 
 	std::string scriptToCall = getApplicationDownloadPath(this->m_AppName.toStdString());
 
-	// StandaloneApp* stlapps = StandaloneApp::GetInstance();
-
-	// stlapps->RetreiveAppSetting(QString::fromStdString(appName));
-	// stlapps->Debug("Function call");
-
-	// if (!(stlapps->GetAction() == "Download" && stlapps->GetStatus() == "Start")) { // if download is not started
-	// 	if (scriptToCall.empty()) { // app not found or delete after extraction
-	// 		stlapps->StoreAppSetting("", "", QString::fromStdString(appName));
-	// 	}
-
-	// 	if (stlapps->GetAction() == "Extract" && stlapps->GetStatus() == "Done") { // if extraction finished
-	// 		scriptToCall = getApplicationDownloadPath(appName);
-
-	// 		stlapps->RetreiveAppSetting(QString::fromStdString(appName));
-	// 		stlapps->Debug("Path Set");
-
-	// 		return scriptToCall;
-	// 	}
-	// 	else if (stlapps->GetAction() == "Extract" && stlapps->GetStatus() == "Start") { // if extraction finished
-	// 		ShowErrorMessage("The application is being installed");
-	// 		updateProgress(50, "Extracting " + appName);
-	// 		return "";
-	// 	}
-	// 	else if (!(stlapps->GetAction() == "Download" && stlapps->GetStatus() == "Done")) { // if download is never started or not done before
 	if (scriptToCall.empty()) {
 		appDownload();
 		
 		return "";
 	}
+	else {
+		ApplicationPreferences::GetInstance()->DeSerializePreferences();
+		bool extractionStarted = QVariant(ApplicationPreferences::GetInstance()->GetLibraExtractionStartedStatus()).toBool();
+		bool extractionFinished = QVariant(ApplicationPreferences::GetInstance()->GetLibraExtractionFinishedStatus()).toBool();
+
+		ApplicationPreferences::GetInstance()->DisplayPreferences();
+
+		if(extractionStarted && !extractionFinished)
+		{
+			QMessageBox::information(this,tr("Extract"),"Extraction in progress");
+			this->close();
+
+			return "";
+		}
+	}
 
 	return scriptToCall;
-	// 	} 
-	// } 
-	// else { // download already started
-	// 	ShowErrorMessage("The application is being downloaded");
-	// 	return "";
-	// }
 }
 
 void StandaloneApp::appDownload()
@@ -94,48 +63,17 @@ void StandaloneApp::appDownload()
 
 	std::string downloadLink = m_appDownloadConfigs["apps"][this->m_AppName.toStdString()][linkyml].as<std::string>();
 
-	// ShowErrorMessage(downloadLink);
-
 	appDownloadDialog.SetPaths(downloadFolder);
 	appDownloadDialog.SetDownloadLink(downloadLink);
 	appDownloadDialog.exec();
 
-	connect( &appDownloadDialog, SIGNAL(doneDownload(QString, QString)), this, SLOT(startUnzip(QString, QString))); 
-	// connect( &appDownloadDialog, SIGNAL(startDownload()), this, SLOT(startDownload()));    
-	// connect( &appDownloadDialog, SIGNAL(cancelDownload()), this, SLOT(cancelDownload()));    
-}
-
-void StandaloneApp::startDownload() 
-{
-	// StandaloneApp* stlapps = StandaloneApp::GetInstance();
-
-	// stlapps->RetreiveAppSetting(appName);
-	// stlapps->Debug("Download Start");
-
-	// stlapps->StoreAppSetting("Download", "Start", appName);
-}
-
-void StandaloneApp::cancelDownload() 
-{
-	// StandaloneApp* stlapps = StandaloneApp::GetInstance();
-
-	// stlapps->RetreiveAppSetting(appName);
-	// stlapps->Debug("Cancel Download");
-
-	// stlapps->StoreAppSetting("", "", appName);
+	connect( &appDownloadDialog, SIGNAL(doneDownload(QString, QString)), this, SLOT(startUnzip(QString, QString)));   
 }
 
 void StandaloneApp::startUnzip(QString fullPath, QString extractPath) 
 {
 	if (cbica::isFile(fullPath.toStdString())) {
-		//  StandaloneApp* stlapps = StandaloneApp::GetInstance();
 
-		//  stlapps->RetreiveAppSetting(appName);
-		//  stlapps->Debug("Done download");
-
-		//  stlapps->StoreAppSetting("Download", "Done", appName);
-
-		//  updateProgress(50, "Extracting " + appName.toStdString());
 		ASyncExtract* asyncExtract = new ASyncExtract();
 
 		connect(asyncExtract, SIGNAL(resultReady(QString)), this, SLOT(doneUnzip()));
@@ -150,11 +88,8 @@ void StandaloneApp::startUnzip(QString fullPath, QString extractPath)
 }
 
 void StandaloneApp::doneUnzip() {
-	//   StandaloneApp* stlapps = StandaloneApp::GetInstance();
 
 	if (getApplicationDownloadPath(this->m_AppName.toStdString()).empty()) {
-
-		// updateProgress(0, "Extracting " + this->m_AppName.toStdString() + " failed");
 
 		QMessageBox::information(NULL,tr("Extraction"),"Extraction failed");
 		// qDebug() << "Extraction failed" << endl;
@@ -166,7 +101,6 @@ void StandaloneApp::doneUnzip() {
 		ApplicationPreferences::GetInstance()->DisplayPreferences();
 	}
 	else {
-		// updateProgress(100, "Extracting " + this->m_AppName.toStdString() + " done");
 		QMessageBox::information(NULL, tr("Extraction"),"Extraction done");
 		// qDebug() << "Extraction done" << endl;
