+++ src/view/gui/StandaloneApp.h
@@ -30,78 +30,71 @@ private:
 
 	Q_DISABLE_COPY(StandaloneApp)
 
-	//! ivars
-	// static StandaloneApp* m_Instance;
-	// static QMutex m_Mutex;
-
 	fAppDownloadDialog appDownloadDialog;
 	QString m_AppName;
 	
 	YAML::Node m_appDownloadConfigs = YAML::LoadFile(getCaPTkDataDir() + "/appsDownloadConfigs.yaml");
 
 	void appDownload();
-	void startDownload();
-	void cancelDownload();
-	
 
 private slots:
 	void startUnzip(QString fullPath, QString extractPath);
 	void doneUnzip();
 };
 
-class ASyncExtract : public QThread
-{
-	Q_OBJECT
-	void run() override {
-		ApplicationPreferences::GetInstance()->SetLibraExtractionStartedStatus(QVariant("true").toString());
-		ApplicationPreferences::GetInstance()->SerializePreferences();
-		ApplicationPreferences::GetInstance()->DisplayPreferences();
+// class ASyncExtract : public QThread
+// {
+// 	Q_OBJECT
+// 	void run() override {
+// 		ApplicationPreferences::GetInstance()->SetLibraExtractionStartedStatus(QVariant("true").toString());
+// 		ApplicationPreferences::GetInstance()->SerializePreferences();
+// 		ApplicationPreferences::GetInstance()->DisplayPreferences();
 
-		if (QFile::exists(this->fullPath))
-		{
-			QZipReader zr(this->fullPath);
-			bool ret = zr.extractAll(this->extractPath);
+// 		if (QFile::exists(this->fullPath))
+// 		{
+// 			QZipReader zr(this->fullPath);
+// 			bool ret = zr.extractAll(this->extractPath);
 
-			if(ret)
-			{
-				//after extraction remove the zip
-				bool successfullyremoved = QFile::remove(this->fullPath.toStdString().c_str());
-			}
-		}
+// 			if(ret)
+// 			{
+// 				//after extraction remove the zip
+// 				bool successfullyremoved = QFile::remove(this->fullPath.toStdString().c_str());
+// 			}
+// 		}
 
-		qDebug() << "Extraction done in background" << this->fullPath << endl;
+// 		qDebug() << "Extraction done in background" << this->fullPath << endl;
 
-		//serialize only once
-		ApplicationPreferences::GetInstance()->SerializePreferences();
+// 		//serialize only once
+// 		ApplicationPreferences::GetInstance()->SerializePreferences();
 
-		emit resultReady(this->appName);
-	}	
+// 		emit resultReady(this->appName);
+// 	}	
 
-public:
-	ASyncExtract() = default;
-	~ASyncExtract() = default;
+// public:
+// 	ASyncExtract() = default;
+// 	~ASyncExtract() = default;
 
-	void setFullPath(QString fullPath) {
-		this->fullPath = fullPath;
-	}
+// 	void setFullPath(QString fullPath) {
+// 		this->fullPath = fullPath;
+// 	}
 
-	void setExtractPath(QString extractPath) {
-		this->extractPath = extractPath;
-	}
+// 	void setExtractPath(QString extractPath) {
+// 		this->extractPath = extractPath;
+// 	}
 
-	void setAppName(QString appName) {
-		this->appName = appName;
-	}
+// 	void setAppName(QString appName) {
+// 		this->appName = appName;
+// 	}
 
-private:
-	Q_DISABLE_COPY(ASyncExtract)
+// private:
+// 	Q_DISABLE_COPY(ASyncExtract)
 
-	QString fullPath;
-	QString extractPath;
-	QString appName;
+// 	QString fullPath;
+// 	QString extractPath;
+// 	QString appName;
 
-signals:
-    void resultReady(QString appName);
-};
+// signals:
+//     void resultReady(QString appName);
+// };
 
 #endif
